Fixed signal handlers and cleaned up code
