+++ src/worker/main.go
@@ -15,8 +15,8 @@ func main() {
 	etcdClient := getClient()
 	defer etcdClient.Close()
 
+	SignalHandlers(etcdClient, etcdServer, &runningVM)
 	startWatchers(etcdClient, &runningVM)
-  
 
 	time.Sleep(2* time.Second)
   	log.Fatal(<-etcdServer.Err()) //Blocking statement
